+++ app-backend/datamodel/src/main/scala/package.scala
@@ -13,7 +13,7 @@ import io.circe.syntax._
 
 import scala.util._
 
-import java.net.URI
+import java.net.{URI, URLDecoder}
 import java.security.InvalidParameterException
 import java.sql.Timestamp
 import java.time.Instant
@@ -299,5 +299,16 @@ package object datamodel extends JsonCodecs {
         case (user, _) => user.id
       }
     }
+  }
+
+  // Lifted from ProjectDao removeLayerOverview method --
+  // it's not clear what sort of common place this URI parsing logic should live in
+  // so it's duplicated here
+  def uriToBucketAndKey(s: String): (String, String) = {
+    val uri = URI.create(s)
+    val urlPath = uri.getPath()
+    val bucket = URLDecoder.decode(uri.getHost(), "UTF-8")
+    val key = URLDecoder.decode(urlPath.drop(1), "UTF-8")
+    (bucket, key)
   }
 }