+++ app-backend/db/src/test/scala/com/azavea/rf/database/AnnotationProjectDaoSpec.scala
@@ -200,7 +200,7 @@ class AnnotationProjectDaoSpec
           )
 
           assert(
-            afterUpdate.ready == annotationProjectUpdate.ready,
+            afterUpdate.status == annotationProjectUpdate.status,
             "Readiness was updated")
 
           true
@@ -220,7 +220,7 @@ class AnnotationProjectDaoSpec
             user <- UserDao.create(userCreate)
             inserted <- AnnotationProjectDao
               .insert(annotationProjectCreate, user)
-            deleted <- AnnotationProjectDao.deleteById(inserted.id)
+            deleted <- AnnotationProjectDao.deleteById(inserted.id, user)
             fetched <- AnnotationProjectDao.getById(inserted.id)
           } yield { (deleted, fetched) }